+++ src/lib/components/validEmails.ts
@@ -1,6 +1,0 @@
-export default function validEmails(text: string) {
-	return text
-		.split(',')
-		.map((email) => email.trim())
-		.every((email) => /.+@.+\..+/.test(email));
-}