+++ src/lib/components/validURL.ts
@@ -1,5 +1,0 @@
-export default function validURL(text: string) {
-	return /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)/.test(
-		text
-	);
-}