+++ src/lib/validation.ts
@@ -1,0 +1,24 @@
+export function validEmail(text: string) {
+	return /.+@.+\..+/.test(text);
+}
+
+export function validEmails(text: string) {
+	return text
+		.split(',')
+		.map((email) => email.trim())
+		.every((email) => /.+@.+\..+/.test(email));
+}
+
+export function validIdentifier(text: string) {
+	return text.length > 0;
+}
+
+export function validURL(text: string) {
+	return /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)/.test(
+		text
+	);
+}
+
+export function validInteger(text: string) {
+	return /[0-9]+/.test(text);
+}